+++ src/lib.rs
@@ -27,6 +27,7 @@ use gpui::App;
 use rust_embed::RustEmbed;
 
 // Core modules
+pub mod element_id;
 pub mod elements;
 pub mod error;
 pub mod fs;